Migrate invite_3b_claimer tests
